requirements.txt
pdfminer.six==20231228
python-pptx==0.6.23
PyMuPDF==1.24.0
pdf2image==1.17.0
Pillow==10.2.0
google-genai>=1.0.0
